+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 - Upgrade palette@4.13.3; styled-components@4.x
 - Switch to use metaphysics v2 schema - justin/alloy
 - Fixes a bug where artist search results in consignments overlap with search bar - yuki24
+- Fixes a bug where location search in the consignments flow stopped working after MP V2 upgrade - zephraph
 - Adds related works section to Artwork page - ashkan18
 - Allows user to save an artwork from the Artwork view - sweir27